[write_ts_table] Remove warning when overwriting files
with 'replace.file' set to TRUE, check whether files
exists before removal: this avoids the warning about
non-existing files

+++ R/functions.R
@@ -73,7 +73,7 @@ write_ts_table <- function(ts, dir, file,
                      file
                  else
                      file.path(dir, file)
-        if (replace.file) {
+        if (replace.file && file.exists(dfile)) {
             file.remove(dfile)
         }
         if (ans == 0L) {
